Cleanup files. Move everything to payment methods
