+++ client/me/payment-methods/payment-method-delete-dialog.tsx
@@ -9,7 +9,6 @@ import { useTranslate, TranslateResult } from 'i18n-calypso';
  */
 import { Button, Dialog } from '@automattic/components';
 import CardHeading from 'calypso/components/card-heading';
-import './style.scss';
 
 interface Props {
 	paymentMethodSummary: TranslateResult;
