--- client/me/purchases/credit-cards/credit-card.scss
@@ -1,4 +1,0 @@
-.credit-card {
-	margin-bottom: 0;
-	display: flex;
-}
