--- client/me/purchases/credit-cards/payment-method-delete-dialog/style.scss
@@ -1,9 +1,0 @@
-.payment-method-delete-dialog {
-	&.dialog {
-		max-width: 440px;
-	}
-
-	.card-heading {
-		margin-top: 0;
-	}
-}
