+++ client/my-sites/purchases/payment-methods/index.tsx
@@ -16,7 +16,7 @@ import DocumentHead from 'calypso/components/data/document-head';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import FormattedHeader from 'calypso/components/formatted-header';
 import PurchasesNavigation from 'calypso/my-sites/purchases/navigation';
-import CreditCards from 'calypso/me/purchases/credit-cards';
+import CreditCards from 'calypso/me/payment-methods/credit-cards';
 import HeaderCake from 'calypso/components/header-cake';
 import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
 import { getAddNewPaymentMethod, getPaymentMethodsUrlFor } from '../paths';
